backend/requirements.txt
Django
djangorestframework
pillow
django-rest-framework-simplejwt
python-dotenv
setuptools
razorpay
environ